+++ hvcc/generators/c2meta/templates/plugin.cpp
@@ -1,0 +1,5 @@
+void init_{{name|lower}}();
+
+extern "C" void init() {
+  init_{{name|lower}}();
+}