+++ hvcc/types/meta.py
@@ -9,6 +9,7 @@ from typing_extensions import Self
 from pydantic import BaseModel, HttpUrl, model_validator
 
 from hvcc.types.GUI import Theme
+from hvcc.generators.c2meta.meta_types import Assets
 
 
 DaisyBoards = Literal['pod', 'petal', 'patch', 'patch_init', 'field']
@@ -83,9 +84,30 @@ class DPF(BaseModel):
         return self
 
 
+class metaPlugin(BaseModel):
+    slug:        str = "module-slug"
+    name:        str = "Module Name"
+    description: str = "Description of the module"
+    tags:        list[str] = []
+    assets:      Optional[Assets] = None
+
+
+class MetaModule(BaseModel):
+    sdk_path:    Optional[str] = None
+    slug:        str = "plugin-slug"
+    name:        str = "Plugin Name"
+    version:     str = "1.0.0"
+    license:     str = "GPL-3.0-only"
+    brand:       str = "Heavy"
+    author:      str = "Heavy"
+    description: str = "Description of the plugin"
+    modules:     list[metaPlugin] = [metaPlugin()]
+
+
 class Meta(BaseModel):
-    name: Optional[str] = None
-    nosimd: Optional[bool] = False
-    daisy: Daisy = Daisy()
-    dpf: DPF = DPF()
+    name:     Optional[str] = None
+    nosimd:   Optional[bool] = False
+    daisy:    Daisy = Daisy()
+    dpf:      DPF = DPF()
+    meta:     MetaModule = MetaModule()
     external: Optional[Dict] = None